shared/grpc_client/client_factory.py
"""
gRPC Client Factory
Creates and manages gRPC client connections to microservices
"""

from typing import Optional, Any
import grpc


class GRPCClientFactory:
    """
    Factory for creating gRPC client stubs with connection pooling
    """
    
    def __init__(self):
        """Initialize the client factory with connection pool"""
        self._channels = {}
        self._stubs = {}
    
    def get_ingestion_client(self, host: str, port: int):
        """
        Get or create an Ingestion Service client
        
        Args:
            host: Service host
            port: Service port
            
        Returns:
            IngestionServiceStub
        """
        pass
    
    def get_analytics_client(self, host: str, port: int):
        """
        Get or create an Analytics Service client
        
        Args:
            host: Service host
            port: Service port
            
        Returns:
            AnalyticsServiceStub
        """
        pass
    
    def get_storage_client(self, host: str, port: int):
        """
        Get or create a Storage Service client
        
        Args:
            host: Service host
            port: Service port
            
        Returns:
            StorageServiceStub
        """
        pass
    
    def get_notification_client(self, host: str, port: int):
        """
        Get or create a Notification Service client
        
        Args:
            host: Service host
            port: Service port
            
        Returns:
            NotificationServiceStub
        """
        pass
    
    def _create_channel(self, host: str, port: int) -> grpc.Channel:
        """
        Create a gRPC channel with connection options
        
        Args:
            host: Service host
            port: Service port
            
        Returns:
            gRPC Channel
        """
        pass
    
    def close_all(self):
        """Close all open gRPC channels"""
        pass


class GRPCInterceptor(grpc.UnaryUnaryClientInterceptor):
    """
    gRPC client interceptor for adding metadata (auth, tracing, etc.)
    """
    
    def __init__(self, tenant_context: Optional[Any] = None):
        """
        Initialize interceptor
        
        Args:
            tenant_context: Optional tenant context for multi-tenancy
        """
        self.tenant_context = tenant_context
    
    def intercept_unary_unary(self, continuation, client_call_details, request):
        """
        Intercept unary-unary calls to add metadata
        """
        pass


# TODO: Implement gRPC client factory with connection pooling
# TODO: Add automatic retry logic for failed requests
# TODO: Add circuit breaker pattern for fault tolerance
# TODO: Add request/response logging
# TODO: Add authentication metadata injection

